DOC: mention MacPorts OpenSSL as option
The failure to use the MacPorts OpenSSL reported in
the message of commit
    621f3f30357d9a4f761a9b2b17baa7aba71d1477
was likely due to an old version of OpenSSL.

Also, the editing of Python's `setup.py` for
configuring the OpenSSL paths was probably due to
using Python 3.6. In Python 3.7 and later, the script
`configure` has an option `--with-openssl` to
define the path to OpenSSL.

Python 3.10 appears to detect and use the MacPorts
OpenSSL without passing `--with-openssl` when building.
The reason appears to be that the script `configure`
looks for `pkg-config`, and asks it for the paths
relevant to OpenSSL. Only one installation of
`pkg-config` exists on my macOS, by MacPorts.

A similar behavior is observed for Python 3.7, 3.8,
and 3.9.

About installations of OpenSSL from source to
`/usr/local/`, I do not remember whether Python 3.6
detected and used that OpenSSL installation
automatically, or I had to configure the OpenSSL
paths in Python's `setup.py` script.